Add save modal to provider page
